requirements.txt
#cv2
opencv-python
mediapipe
numpy
cvzone
PyQt6
pyyaml